wp-content/themes/blocksy-sportsmed/.bak/backup/mybusiness_api/requirements.txt

Flask==2.0.2
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.26.1
Authlib==0.15.5
gunicorn==20.1.0

